standaloneusers - split User.PasswordResetEmail into public and private actions
