+++ ext/standaloneusers/ang/crmResetPassword/crmResetPassword.html
@@ -16,7 +16,7 @@
           />
     </div>
 
-    <button class=crm-button ng-click="$ctrl.sendPasswordReset()" >{{ts('Send Password Reset')}}</button>
+    <button class=crm-button ng-click="$ctrl.requestPasswordResetEmail()" >{{ts('Request Password Reset')}}</button>
   </form>
 
   <div ng-if="$ctrl.resetSuccessfullySubmitted" class="help" >
